requirements.txt
scrapy
requests
requests_oauthlib
pyyaml
